Extract LaunchProfileProperties
Extract `LaunchProfileProperties` into its own top-level class. This will make testing somewhat easier, simplifies `LaunchProfileProjectPropertiesProvider`, and is more in line with how we structure code elsewhere in the repo.
